Refactor in pkg/web_app: focusability method separated from the rest of modal window handling.
